Add types.ts file
